fix: add platform specific edge cases and tests
