+++ packages/caption/__tests__/caption.ios.native.test.js
@@ -1,0 +1,15 @@
+/* eslint-env jest */
+
+import "react-native";
+import shared from "./shared";
+
+jest.mock("react-native", () => {
+  const reactNative = require.requireActual("react-native");
+  reactNative.Platform.OS = "ios";
+  jest
+    .spyOn(reactNative.Platform, "select")
+    .mockImplementation(obj => obj.ios || obj.default);
+  return reactNative;
+});
+
+shared();
